+++ test/cases/escape-method-render/webpack.config.js
@@ -26,7 +26,6 @@ module.exports = {
         },
       },
 
-      // process required images in pug
       {
         test: /\.(png|jpg|jpeg)/,
         type: 'asset/resource',