+++ test/cases/exception-pug-compile/webpack.config.js
@@ -1,6 +1,3 @@
-const path = require('path');
-const basePath = path.resolve(__dirname);
-
 module.exports = {
   mode: 'production',
   entry: {