+++ test/cases/javascript-esmodule-false-require/webpack.config.js
@@ -39,7 +39,6 @@ module.exports = {
         },
       },
 
-      // process required images in pug
       {
         test: /\.(png|jpg|jpeg)/,
         type: 'asset/resource',