+++ test/cases/javascript-option-method-html/webpack.config.js
@@ -50,7 +50,6 @@ module.exports = {
         ],
       },
 
-      // process required images in pug
       {
         test: /\.(png|jpg|jpeg)/,
         type: 'asset/resource',