+++ test/cases/require-assets-method-compile/webpack.config.js
@@ -23,11 +23,9 @@ module.exports = {
     rules: [
       {
         test: /\.pug$/,
-        //loader: PugPlugin.loader,
-        loader: 'pug-loader', // local dev
+        loader: 'pug-loader',
         options: {
           method: 'compile',
-          //method: 'render',
         },
       },
 
@@ -35,7 +33,6 @@ module.exports = {
         test: /\.(css|sass|scss)$/,
         type: 'asset/resource',
         generator: {
-          // save required styles
           filename: 'assets/css/[name].[contenthash:8].css',
         },
         use: [
@@ -45,7 +42,6 @@ module.exports = {
         ],
       },
 
-      // process required images in pug
       {
         test: /\.(png|jpg|jpeg)/,
         type: 'asset/resource',