+++ test/cases/require-json-alias/webpack.config.js
@@ -29,7 +29,6 @@ module.exports = {
         test: /\.pug$/,
         loader: 'pug-loader',
         options: {
-          //method: 'compile',
           method: 'render',
         },
       },