+++ test/cases/require-json-relative/webpack.config.js
@@ -22,7 +22,6 @@ module.exports = {
         test: /\.pug$/,
         loader: 'pug-loader',
         options: {
-          //method: 'compile',
           method: 'render',
         },
       },