+++ test/cases/require-resource-variable-parent-dir/webpack.config.js
@@ -36,11 +36,9 @@ module.exports = {
         loader: 'pug-loader',
         options: {
           method: 'compile',
-          //method: 'render',
         },
       },
 
-      // process required images in pug
       {
         test: /\.(png|jpg|jpeg)/,
         type: 'asset/resource',